@@ -1,15 +1,17 @@
 package de.tobiasbell.aoc.solution._2015.day6;
 
+import org.jetbrains.annotations.NotNull;
+
 public enum NewCommand implements LightCommand {
     TurnOn {
         @Override
-        public Brightness evaluate(Brightness currentState) {
+        public @NotNull Brightness evaluate(@NotNull Brightness currentState) {
             return new Brightness(currentState.brightness() + 1);
         }
     },
     TurnOff {
         @Override
-        public Brightness evaluate(Brightness currentState) {
+        public @NotNull Brightness evaluate(@NotNull Brightness currentState) {
             return currentState.equals(Brightness.Off)
                     ? currentState
                     : new Brightness(currentState.brightness() - 1);
@@ -17,12 +19,12 @@ public enum NewCommand implements LightCommand {
     },
     Toggle {
         @Override
-        public Brightness evaluate(Brightness currentState) {
+        public @NotNull Brightness evaluate(@NotNull Brightness currentState) {
             return new Brightness(currentState.brightness() + 2);
         }
     };
 
-    public static LightCommand of(final String command) {
+    public static @NotNull LightCommand of(final @NotNull String command) {
         return switch (command) {
             case "turn on" -> NewCommand.TurnOn;
             case "turn off" -> NewCommand.TurnOff;
