@@ -1,10 +1,12 @@
 package de.tobiasbell.aoc.solution._2015.day7;
 
 
+import org.jetbrains.annotations.Nullable;
+
 public class Wire {
     private static final int MAX_SIGNAL = 65_535;
     private final String name;
-    private Integer signal;
+    private @Nullable Integer signal;
     private Expression inputExpression;
 
     public Wire(String name) {
