@@ -1,5 +1,7 @@
 package de.tobiasbell.aoc.solution._2015.day7;
 
+import org.jetbrains.annotations.NotNull;
+
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -8,11 +10,11 @@ import java.util.Set;
 public class WireTable {
     private final Map<String, Wire> table = new HashMap<>();
 
-    public Wire getWire(final String name) {
+    public @NotNull Wire getWire(final String name) {
         return table.computeIfAbsent(name, Wire::new);
     }
 
-    public Set<String> getWireNames() {
+    public @NotNull Set<String> getWireNames() {
         return Collections.unmodifiableSet(table.keySet());
     }
 
