@@ -3,17 +3,23 @@ package de.tobiasbell.aoc.loader;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.junit5.WireMockRuntimeInfo;
 import com.github.tomakehurst.wiremock.junit5.WireMockTest;
+import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
 
-import java.io.IOException;
-
-import static com.github.tomakehurst.wiremock.client.WireMock.*;
+import static com.github.tomakehurst.wiremock.client.WireMock.anyRequestedFor;
+import static com.github.tomakehurst.wiremock.client.WireMock.anyUrl;
+import static com.github.tomakehurst.wiremock.client.WireMock.getRequestedFor;
+import static com.github.tomakehurst.wiremock.client.WireMock.matching;
+import static com.github.tomakehurst.wiremock.client.WireMock.ok;
+import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
+import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
+import static com.github.tomakehurst.wiremock.client.WireMock.verify;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @WireMockTest
 class WebLoaderTest {
 
-    public static void main(String[] args) throws IOException, InterruptedException {
+    public static void main(String[] args) {
         final String sessionId = "xxx";
         final WebLoader webLoader = WebLoader.withSession(sessionId);
         final String input = webLoader.loadInput(2020, 20);
@@ -21,7 +27,7 @@ class WebLoaderTest {
     }
 
     @Test
-    void callsCorrectUrl(WireMockRuntimeInfo wmRuntimeInfo) throws IOException, InterruptedException {
+    void callsCorrectUrl(@NotNull WireMockRuntimeInfo wmRuntimeInfo) {
         // given
         final WebLoader webLoader = WebLoader.withBaseUrl(wmRuntimeInfo.getHttpBaseUrl(), "sessionId");
         // when
@@ -31,11 +37,10 @@ class WebLoaderTest {
     }
 
     @Test
-    void downloadsData(WireMockRuntimeInfo wmRuntimeInfo) throws IOException, InterruptedException {
+    void downloadsData(@NotNull WireMockRuntimeInfo wmRuntimeInfo) {
         // given
-        final WireMock wireMock = wmRuntimeInfo.getWireMock();
         final WebLoader webLoader = WebLoader.withBaseUrl(wmRuntimeInfo.getHttpBaseUrl(), "sessionId");
-        wireMock.stubFor(WireMock.get(urlEqualTo("/2015/day/1/input"))
+        stubFor(WireMock.get(urlEqualTo("/2015/day/1/input"))
                 .willReturn(ok("input")));
         // when
         final String result = webLoader.loadInput(2015, 1);
@@ -44,14 +49,13 @@ class WebLoaderTest {
     }
 
     @Test
-    void sendsAuthenticationCookie(WireMockRuntimeInfo wmRuntimeInfo) throws IOException, InterruptedException {
+    void sendsAuthenticationCookie(@NotNull WireMockRuntimeInfo wmRuntimeInfo) {
         // given
-        final WireMock wireMock = wmRuntimeInfo.getWireMock();
         final WebLoader webLoader = WebLoader.withBaseUrl(wmRuntimeInfo.getHttpBaseUrl(), "sessionId");
         // when
         webLoader.loadInput(2015, 1);
         //then
-        wireMock.verify(anyRequestedFor(anyUrl())
+        verify(anyRequestedFor(anyUrl())
                 .withCookie("session", matching("sessionId")));
     }
 }
